# Don't get the dispatcher when the cached interest is NEVER