--- a/tokio-trace/Cargo.toml
+++ b/tokio-trace/Cargo.toml
@@ -14,6 +14,7 @@ tokio-trace-core = { path = "../tokio-trace-core" }
 ansi_term = "0.11"
 humantime = "1.1.1"
 env_logger = "0.5"
+log = "0.4"
 tokio-trace-log = { path = "../tokio-trace-log" }
 tokio-trace-core = { path = "../tokio-trace-core", features = ["test-support"] }
 futures = "0.1"