Currently, the `Callsite::is_enabled` function requires a reference to a
dispatcher. This is used to filter the callsite when the cached interest
is `Interest::SOMETIMES`. However, it means we must always get the
current dispatcher, even when the interest is `NEVER`.

This branch replaces `Callsite::is_enabled` with a `Callsite::interest`
function that just returns the cached interest. This no longer requires
a reference to a dispatcher. The `Span::new` and `Event::observe`
functions have been rewritten to only get the dispatcher if the interest
is `ALWAYS` or `SOMETIMES`.

This results in a major performance improvement in the case where a span
is disabled, putting our performance within the same order of magnitude
as the `log` crate's performance when there's no logger.

Benchmark results  (`bench_span_no_subscriber`) is the most relevant:

```
     Running target/release/deps/no_subscriber-6fb6999bbb120c50

running 5 tests
test bench_1_atomic_load              ... bench:           0 ns/iter (+/- 0)
test bench_costly_field_no_subscriber ... bench:           5 ns/iter (+/- 1)
test bench_log_no_logger              ... bench:           1 ns/iter (+/- 0)
test bench_no_span_no_subscriber      ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber         ... bench:           5 ns/iter (+/- 2)

test result: ok. 0 passed; 0 failed; 0 ignored; 5 measured; 0 filtered out

     Running target/release/deps/subscriber-c1100adcfb21cd44

running 4 tests
test span_no_fields            ... bench:          68 ns/iter (+/- 10)
test span_repeatedly           ... bench:       7,877 ns/iter (+/- 1,093)
test span_with_fields          ... bench:          86 ns/iter (+/- 9)
test span_with_fields_add_data ... bench:         770 ns/iter (+/- 104)
```

vs. master:

```
     Running target/release/deps/no_subscriber-9f5296229e2944c8

running 4 tests
test bench_1_atomic_load              ... bench:           1 ns/iter (+/- 0)
test bench_costly_field_no_subscriber ... bench:          42 ns/iter (+/- 7)
test bench_no_span_no_subscriber      ... bench:           0 ns/iter (+/- 0)
test bench_span_no_subscriber         ... bench:          39 ns/iter (+/- 7)

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out

     Running target/release/deps/subscriber-a13cd106471c1486

running 4 tests
test span_no_fields            ... bench:          77 ns/iter (+/- 14)
test span_repeatedly           ... bench:       9,305 ns/iter (+/- 2,155)
test span_with_fields          ... bench:          92 ns/iter (+/- 16)
test span_with_fields_add_data ... bench:         814 ns/iter (+/- 136)

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out

```

Signed-off-by: Eliza Weisman <eliza@buoyant.io>